requirements.txt
Flask==3.1.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
python-dotenv==1.0.1
SQLAlchemy==2.0.27
pandas==2.2.3
numpy==2.0.2
plotly==5.18.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.3.0